Remove unused babel_route pointer from convert_from_rte
Initially, it was planned to use the route pointer in the parameters to
apply the modified route attributes (currently only metric).

However, the existing implementation does not allow this to be done
easily, so instead the route attributes are handed over using explicit
pointers as function parameters.

Remove the unused function parameter